Migrated MsSqlDatabaseTester constructor methods
