+++ tests/dbSpecflow_tests/dbSpecflow_tests.csproj
@@ -174,6 +174,7 @@
     <Compile Include="MsSqlDatabaseTester\HasSchema_Tests.cs" />
     <Compile Include="MsSqlDatabaseTester\ExecuteStatementWithoutResult_Tests.cs" />
     <Compile Include="MsSqlDatabaseTester\HasTable_Tests.cs" />
+    <Compile Include="MsSqlDatabaseTester\NewObject_Tests.cs" />
     <Compile Include="MsSqlDatabaseTester\RowCount_Tests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="DatabaseTesterTests.cs" />
